here we go!
there are many many bugs, I did not get as much implemented as I wanted to get done. I spent the majority of my time this evening rewriting code, and debugging that code. I am still in the process of implementing a "add deposit" function. As soon as I successfully do that, I feel that the "withdrawl" function would work easily as well. My brain hurts. I wish I was a savant, but alas, I am not.